move: UserEntity.kt 파일 위치 변경
core.user.domain > core.user.database 로 위치 변경
