+++ src/main/kotlin/com/wafflestudio/toyproject/team4/core/user/api/request/RegisterRequest.kt
@@ -1,6 +1,6 @@
 package com.wafflestudio.toyproject.team4.core.user.api.request
 
-import com.wafflestudio.toyproject.team4.core.user.domain.UserEntity
+import com.wafflestudio.toyproject.team4.core.user.database.UserEntity
 
 data class RegisterRequest(
     val username: String,
